feat: remove name field, use currentUser prop from connected account
