+++ apps/frontend/src/routes/MovieDetailPage.tsx
@@ -1,9 +1,8 @@
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import { useState } from "react";
 import { ArrowLeft } from "lucide-react";
 import MovieHero from "@/components/ui/MovieHero";
 import RateMovie from "@/components/ui/RateMovie";
-import ReviewCard from "@/components/ui/CommentSectionComponent";
+import CommentSection from "@/components/ui/CommentSectionComponent";
 
 export const Route = createFileRoute("/MovieDetailPage")({
   component: MovieDetailPage,
@@ -20,14 +19,6 @@ interface Film {
   synopsis: string;
 }
 
-interface MovieComment {
-  id: number;
-  username: string;
-  date: string;
-  reviewText: string;
-  rating: number;
-}
-
 // données mockées du film
 const mockMovie: Film = {
   title: "Inception",
@@ -40,33 +31,11 @@ const mockMovie: Film = {
     "Dom Cobb est un voleur spécialisé dans l'art de s'introduire dans les rêves des autres pour leur subtiliser les secrets de leur subconscient. Ce talent rare en a fait un joueur très recherché dans le monde trouble de l'espionnage industriel. Mais cette activité lui a coûté cher : il a perdu tout ce qu'il aimait. On lui offre une chance de se racheter : accomplir une mission en apparence impossible, l'inception.",
 };
 
+// utilisateur connecté mocké
+const currentUser = "Julie Caty";
+
 function MovieDetailPage() {
   const navigate = useNavigate();
-
-  const [comments, setComments] = useState<MovieComment[]>([]);
-  const [showForm, setShowForm] = useState(false);
-  const [formUsername, setFormUsername] = useState("");
-  const [formText, setFormText] = useState("");
-  const [formRating, setFormRating] = useState(0);
-  const [hoveredStar, setHoveredStar] = useState(0);
-
-  function handleSubmit() {
-    if (!formUsername.trim() || !formText.trim() || formRating === 0) return;
-
-    const newComment: MovieComment = {
-      id: Date.now(),
-      username: formUsername.trim(),
-      date: "À l'instant",
-      reviewText: formText.trim(),
-      rating: formRating,
-    };
-
-    setComments([...comments, newComment]);
-    setFormUsername("");
-    setFormText("");
-    setFormRating(0);
-    setShowForm(false);
-  }
 
   return (
     <div className="min-h-screen bg-gray-950">
@@ -110,68 +79,8 @@ function MovieDetailPage() {
         </section>
 
         {/* Section Commentaires */}
-        <section className="px-4 space-y-4">
-          <div className="flex items-center justify-between">
-            <h2 className="text-white font-bold text-lg">Commentaires</h2>
-            <button
-              onClick={() => setShowForm(!showForm)}
-              className="bg-red-600 text-white text-sm px-4 py-2 rounded-lg hover:bg-red-700 transition-colors"
-            >
-              {showForm ? "Annuler" : "Ajouter un commentaire"}
-            </button>
-          </div>
-
-          {showForm && (
-            <div className="bg-gray-800 border border-gray-700 rounded-2xl p-4 flex flex-col gap-3">
-              <input
-                type="text"
-                placeholder="Votre nom"
-                value={formUsername}
-                onChange={(e) => setFormUsername(e.target.value)}
-                className="bg-gray-900 border border-gray-600 text-white text-sm rounded-lg px-3 py-2 outline-none focus:border-red-500 transition-colors"
-              />
-              <textarea
-                placeholder="Votre commentaire..."
-                value={formText}
-                onChange={(e) => setFormText(e.target.value)}
-                rows={3}
-                className="bg-gray-900 border border-gray-600 text-white text-sm rounded-lg px-3 py-2 outline-none focus:border-red-500 transition-colors resize-none"
-              />
-              <div className="flex items-center gap-1">
-                {[1, 2, 3, 4, 5].map((i) => (
-                  <span
-                    key={i}
-                    className={`text-2xl cursor-pointer transition-colors ${i <= (hoveredStar || formRating) ? "text-yellow-400" : "text-gray-600"}`}
-                    onMouseEnter={() => setHoveredStar(i)}
-                    onMouseLeave={() => setHoveredStar(0)}
-                    onClick={() => setFormRating(i)}
-                  >
-                    ★
-                  </span>
-                ))}
-              </div>
-              <button
-                onClick={handleSubmit}
-                className="bg-red-600 text-white font-bold text-sm px-4 py-2 rounded-lg hover:bg-red-700 transition-colors"
-              >
-                Publier
-              </button>
-            </div>
-          )}
-
-          {comments.length > 0 && (
-            <div className="flex flex-col gap-3">
-              {comments.map((comment) => (
-                <ReviewCard
-                  key={comment.id}
-                  username={comment.username}
-                  date={comment.date}
-                  reviewText={comment.reviewText}
-                  rating={comment.rating}
-                />
-              ))}
-            </div>
-          )}
+        <section className="px-4">
+          <CommentSection currentUser={currentUser} />
         </section>
 
       </div>
